Fix issue #5, empty rules when using webpack
This fixes an issue when using webpack and PostCSS 6,
where the rules being generated inside @media rules would
not contain their properties.

It has something to do with the order of appending props
and inserting the rule.

The previous approach inserted the rule then filled it with props.
The new approach fills the rule with props then inserts it.

Either approach works fine with vanilla PostCSS, the problem is
either with webpack or postcss-loader.